Add files via upload
This is a prototype of a application which can tell you what is your probability of having a heart disease. It is based on supervised machine learning models, to be exact it is based on RandomForest and LogisticRegression. The data used in learning process is from kaggle.com.
Link to dataset: https://www.kaggle.com/datasets/kamilpytlak/personal-key-indicators-of-heart-disease